# frontend/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InvoiceFlow API",
    "description": "E-invoice processing platform with AI extraction and global compliance (PEPPOL, ZATCA, Brazil NFe, India IRP, Mexico CFDI, CTC reporting, and post-audit archival).",
    "version": "0.2.0"
  },
  "servers": [
    { "url": "http://localhost:8000", "description": "Local development" },
    { "url": "/api/v1", "description": "Production (Nginx proxy)" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/v1/auth/login"
      }
    },
    "schemas": {
      "InvoiceStatus": {
        "type": "string",
        "enum": ["received", "processing", "extracted", "reviewing", "approved", "compliant", "transmitted", "failed", "rejected"]
      },
      "ComplianceModel": {
        "type": "string",
        "enum": ["peppol", "clearance", "ctc", "post_audit"]
      },
      "ExtractionMethod": {
        "type": "string",
        "enum": ["ocr", "llm", "xml", "manual"]
      },
      "IngestionSource": {
        "type": "string",
        "enum": ["email", "ftp", "sftp", "api", "webhook", "manual"]
      },
      "InvoiceLine": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "invoice_id": { "type": "string", "format": "uuid" },
          "line_number": { "type": "integer" },
          "description": { "type": "string", "nullable": true },
          "quantity": { "type": "number" },
          "unit_price": { "type": "number" },
          "tax_rate": { "type": "number" },
          "tax_amount": { "type": "number" },
          "line_total": { "type": "number" },
          "item_code": { "type": "string", "nullable": true }
        },
        "required": ["id", "invoice_id", "line_number", "quantity", "unit_price", "tax_rate", "tax_amount", "line_total"]
      },
      "InvoiceLineCreate": {
        "type": "object",
        "properties": {
          "line_number": { "type": "integer" },
          "description": { "type": "string", "nullable": true },
          "quantity": { "type": "number", "default": 1.0 },
          "unit_price": { "type": "number", "default": 0.0 },
          "tax_rate": { "type": "number", "default": 0.0 },
          "item_code": { "type": "string", "nullable": true }
        },
        "required": ["line_number"]
      },
      "Invoice": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string", "format": "uuid" },
          "status": { "$ref": "#/components/schemas/InvoiceStatus" },
          "invoice_number": { "type": "string" },
          "vendor_name": { "type": "string" },
          "vendor_tax_id": { "type": "string", "nullable": true },
          "buyer_name": { "type": "string", "nullable": true },
          "buyer_tax_id": { "type": "string", "nullable": true },
          "invoice_date": { "type": "string", "format": "date-time", "nullable": true },
          "due_date": { "type": "string", "format": "date-time", "nullable": true },
          "currency": { "type": "string" },
          "subtotal": { "type": "number" },
          "tax_amount": { "type": "number" },
          "total_amount": { "type": "number" },
          "country_code": { "type": "string" },
          "compliance_model": { "$ref": "#/components/schemas/ComplianceModel" },
          "source": { "$ref": "#/components/schemas/IngestionSource" },
          "source_reference": { "type": "string", "nullable": true },
          "file_url": { "type": "string", "nullable": true },
          "original_filename": { "type": "string", "nullable": true },
          "ocr_confidence": { "type": "number", "nullable": true },
          "extraction_method": { "$ref": "#/components/schemas/ExtractionMethod" },
          "extracted_data": { "type": "object", "additionalProperties": true },
          "compliance_response": { "type": "object", "additionalProperties": true },
          "error_message": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "lines": { "type": "array", "items": { "$ref": "#/components/schemas/InvoiceLine" } },
          "audit_log": { "type": "array", "items": { "$ref": "#/components/schemas/AuditLog" }, "description": "Audit trail entries (included in detail view)" }
        },
        "required": ["id", "tenant_id", "status", "invoice_number", "vendor_name", "currency", "subtotal", "tax_amount", "total_amount", "country_code", "compliance_model", "source", "created_at", "updated_at"]
      },
      "InvoiceCreate": {
        "type": "object",
        "properties": {
          "invoice_number": { "type": "string", "maxLength": 100 },
          "vendor_name": { "type": "string", "maxLength": 255 },
          "vendor_tax_id": { "type": "string", "nullable": true },
          "buyer_name": { "type": "string", "nullable": true },
          "buyer_tax_id": { "type": "string", "nullable": true },
          "invoice_date": { "type": "string", "format": "date-time", "nullable": true },
          "due_date": { "type": "string", "format": "date-time", "nullable": true },
          "currency": { "type": "string", "default": "USD" },
          "country_code": { "type": "string", "maxLength": 2 },
          "compliance_model": { "$ref": "#/components/schemas/ComplianceModel" },
          "lines": { "type": "array", "items": { "$ref": "#/components/schemas/InvoiceLineCreate" }, "default": [] }
        },
        "required": ["invoice_number", "vendor_name", "country_code", "compliance_model"]
      },
      "InvoiceUpdate": {
        "type": "object",
        "properties": {
          "invoice_number": { "type": "string", "nullable": true },
          "vendor_name": { "type": "string", "nullable": true },
          "vendor_tax_id": { "type": "string", "nullable": true },
          "buyer_name": { "type": "string", "nullable": true },
          "buyer_tax_id": { "type": "string", "nullable": true },
          "invoice_date": { "type": "string", "format": "date-time", "nullable": true },
          "due_date": { "type": "string", "format": "date-time", "nullable": true },
          "currency": { "type": "string", "nullable": true },
          "status": { "$ref": "#/components/schemas/InvoiceStatus" }
        }
      },
      "InvoiceListResponse": {
        "type": "object",
        "properties": {
          "invoices": { "type": "array", "items": { "$ref": "#/components/schemas/Invoice" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" }
        },
        "required": ["invoices", "total", "page", "page_size"]
      },
      "AuditLog": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string", "format": "uuid" },
          "invoice_id": { "type": "string", "format": "uuid", "nullable": true },
          "action": { "type": "string" },
          "details": { "type": "object", "additionalProperties": true },
          "user_id": { "type": "string", "nullable": true },
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "tenant_id", "action", "timestamp"]
      },
...
      "FileUploadResponse": {
        "type": "object",
        "properties": {
          "invoice_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Null when processing is async; use task_id for tracking" },
          "status": { "$ref": "#/components/schemas/InvoiceStatus" },
          "filename": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["status", "filename", "message"]
      },
      "PollResponse": {
        "type": "object",
        "properties": {
          "task_id": { "type": "string" },
          "source_type": { "$ref": "#/components/schemas/IngestionSource" },
          "status": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["task_id", "source_type", "status", "message"]
      },
      "IngestionStatsResponse": {
        "type": "object",
        "properties": {
          "total_ingested": { "type": "integer" },
          "by_source": { "type": "object", "additionalProperties": { "type": "integer" } },
          "last_poll_at": { "type": "string", "format": "date-time", "nullable": true },
          "pending_processing": { "type": "integer" }
        },
        "required": ["total_ingested", "by_source", "pending_processing"]
      },
      "WebhookPayload": {
        "type": "object",
        "properties": {
          "filename": { "type": "string" },
          "file_content_b64": { "type": "string", "description": "Base64-encoded file content" },
          "mime_type": { "type": "string", "default": "application/pdf" },
          "source_reference": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        },
        "required": ["filename", "file_content_b64"]
      },
      "TaskStatusResponse": {
        "type": "object",
        "properties": {
          "task_id": { "type": "string" },
          "status": { "type": "string" },
          "result": { "type": "string", "nullable": true },
          "traceback": { "type": "string", "nullable": true }
        },
        "required": ["task_id", "status"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string" }
        },
        "required": ["detail"]
      }
    }
  },
  "paths": {
...
    "/api/v1/invoices": {
      "get": {
        "tags": ["invoices"],
        "summary": "List invoices with pagination and filters",
        "operationId": "listInvoices",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/InvoiceStatus" } },
          { "name": "country", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search by invoice number or vendor name" }
        ],
        "responses": {
          "200": { "description": "Paginated invoice list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvoiceListResponse" } } } }
        }
      },
      "post": {
        "tags": ["invoices"],
        "summary": "Create a new invoice (manual entry)",
        "operationId": "createInvoice",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvoiceCreate" } } } },
        "responses": {
          "201": { "description": "Invoice created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invoice" } } } }
        }
      }
    },
    "/api/v1/invoices/{invoice_id}": {
      "get": {
        "tags": ["invoices"],
        "summary": "Get invoice by ID",
        "operationId": "getInvoice",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "invoice_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Invoice details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invoice" } } } },
          "404": { "description": "Invoice not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "put": {
        "tags": ["invoices"],
        "summary": "Update invoice",
        "operationId": "updateInvoice",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "invoice_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvoiceUpdate" } } } },
        "responses": {
          "200": { "description": "Invoice updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invoice" } } } }
        }
      },
      "delete": {
        "tags": ["invoices"],
        "summary": "Delete invoice",
        "operationId": "deleteInvoice",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "invoice_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "204": { "description": "Invoice deleted" } }
      }
    },
    "/api/v1/invoices/{invoice_id}/approve": {
      "post": {
        "tags": ["invoices"],
        "summary": "Approve invoice for compliance processing",
        "operationId": "approveInvoice",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "invoice_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Invoice approved", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invoice" } } } }
        }
      }
    },
    "/api/v1/invoices/{invoice_id}/reject": {
      "post": {
        "tags": ["invoices"],
        "summary": "Reject invoice",
        "operationId": "rejectInvoice",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "invoice_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Invoice rejected", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invoice" } } } }
        }
      }
    },
    "/api/v1/ingestion/upload": {
      "post": {
        "tags": ["ingestion"],
        "summary": "Upload an invoice file for processing",
        "operationId": "uploadInvoice",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary", "description": "Invoice file (PDF, XML, PNG, JPEG, TIFF)" } }, "required": ["file"] } } } },
        "responses": {
          "200": { "description": "File uploaded and queued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileUploadResponse" } } } }
        }
      }
    },
    "/api/v1/ingestion/webhook": {
      "post": {
        "tags": ["ingestion"],
        "summary": "Receive invoice via webhook",
        "operationId": "webhookReceive",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookPayload" } } } },
        "responses": {
          "200": { "description": "Webhook received", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileUploadResponse" } } } }
        }
      }
    },
    "/api/v1/ingestion/poll/email": {
      "post": {
        "tags": ["ingestion"],
        "summary": "Trigger email inbox poll",
        "operationId": "triggerEmailPoll",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Poll initiated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PollResponse" } } } }
        }
      }
    },
    "/api/v1/ingestion/poll/ftp": {
      "post": {
        "tags": ["ingestion"],
        "summary": "Trigger FTP folder poll",
        "operationId": "triggerFtpPoll",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Poll initiated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PollResponse" } } } }
        }
      }
    },
    "/api/v1/ingestion/poll/sftp": {
      "post": {
        "tags": ["ingestion"],
        "summary": "Trigger SFTP folder poll",
        "operationId": "triggerSftpPoll",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Poll initiated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PollResponse" } } } }
        }
      }
    },
    "/api/v1/ingestion/stats": {
      "get": {
        "tags": ["ingestion"],
        "summary": "Get ingestion statistics",
        "operationId": "getIngestionStats",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Ingestion statistics", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IngestionStatsResponse" } } } }
        }
      }
    },
    "/api/v1/ingestion/task/{task_id}": {
      "get": {
        "tags": ["ingestion"],
        "summary": "Check ingestion task status",
        "operationId": "getIngestionTaskStatus",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "task_id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Task status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TaskStatusResponse" } } } }
        }
      }
    },
...
  },
  "security": [{ "BearerAuth": [] }]
}